feat: standardize caching on the resource selector
