+++ models/config.go
@@ -92,8 +92,8 @@ type ConfigItem struct {
 	DeleteReason  string               `json:"delete_reason,omitempty"`
 }
 
-func (c ConfigItem) Key() string {
-	return c.ID.String()
+func (ConfigItem) TableName() string {
+	return "config_items"
 }
 
 func (ci *ConfigItem) SetParent(parent *ConfigItem) {
